check for apt-cyg and skip install if it's already there. wget/https comment. improve doc.

+++ _setup/_installDependencies.sh
@@ -1,3 +1,9 @@
+# DESCRIPTION
+# Installs Windows dependencies for this autobrood repository's scripts.
+
+# DEPENDENCIES
+# wget with https capability (since now everyone in the universe is encrypting all web traffice even if needlessly), which you can get and get in your path via my _ebSuperBin and _ebPathMan repositories. Cygwin.
+
 # NOTES
 # For at least the chere -i -t mintty command to work, you may need to run this from a cygwin prompt started as a Windows Administrator.
 # If install fails (canna download from mirror), try:
@@ -7,6 +13,7 @@
 # apt-cyg mirror "ftp://mirrors.syringanetworks.net/cygwin/"
 # re: http://superuser.com/a/941135
 # ALSO, mysteriously, installing perl xml_twig via cygwin may solve any errors after this script successfully runs all commands. 
+# PREVIOUS DOWNLOAD COMMANDS used: curl -LO <that-web-address>
 
 # TO DO: apt-cyg check (which apt-cyg) and auto-install of perl and ~libxml, wget.
 
@@ -19,19 +26,21 @@ if [ "$CONDITION" == "y" ]; then
 				echo Trying some stuff to install dependencies for autobrood using cygwin . . .
 				# cat ./my.minttyrc.settings.txt > /home/$username/.minttyrc
 		# Re goodies divulged here: http://stackoverflow.com/q/9260014/1397555
-		# wget --no-check-certificate raw.github.com/transcode-open/apt-cyg/master/apt-cyg
-		curl https://raw.githubusercontent.com/transcode-open/apt-cyg/master/apt-cyg > apt-cyg
-		chmod +x apt-cyg
-		mv apt-cyg /usr/local/bin
-		cd /usr/local/bin
-			# Optional but preferred:
-			apt-cyg install chere
-			chere -i -t mintty
+		# Only install apt-cyg if it is not installed:
+		if [ ! -e /usr/local/bin/apt-cyg ]
+		then
+			wget https://raw.github.com/transcode-open/apt-cyg/master/apt-cyg > apt-cyg
+			chmod +x apt-cyg
+			mv apt-cyg /usr/local/bin
+		fi
+		# Optional but preferred:
+		apt-cyg install chere
+		chere -i -t mintty
 		apt-cyg install make
 # IF YOU comment out the next two lines, make sure to install these modules via the Cygwin installer! :
 		apt-cyg install perl
 		apt-cyg install perl-libxml-perl
-		curl -LO http://www.xmltwig.org/xmltwig/XML-Twig-3.50.tar.gz
+		wget https://www.xmltwig.org/xmltwig/XML-Twig-3.50.tar.gz
 		tar xz < XML-Twig-3.50.tar.gz
 		cd XML-Twig-3.50
 		perl Makefile.PL -y
